requirements.txt
tweepy
python-dotenv
kafka-python
langdetect
pyspark
findspark
nzpyida
streamlit
wordcloud
enchant
nltk
numpy
pandas
plotly
matplotlib